Copy-friendly config definition
The purpose of this commit is to refactor how we define the location of
the `config.json` to mimic how the `to` property of a `copy` command
works. This has several advantages:

- Less need to transform the property to match what a `copy` command
expects (this is more noticeable in the case of the `edison`).
- Reuse the way we describe a location in a partition instead of
reinventing things.
- Move the knowledge of how a config file should be named to the specs,
instead of hardcoding it in the client.